Added real time updating to chart creation. More details below
There were a few changes to the file managment in this commit.
makeNew.php now only contains the html and the added js/makenew.js
contains all the ajax/jquery code needed for the chart making process.

I did not work on making the charts work for two companies; that will
come soon.